Implement PyArrow schemas in Pydantic metadata classes
Rather than hard-coding the EPA CEMS PyArrow schema for outputting the parquet files,
dynamically generate it using our metadata store.

In the process, I discovered that we had been casting `unit_id_epa` to Int32, even
though it's actually a string. Not sure what the consequences of that are, and it seems
weird that it never complained, as there are definitely letters in there. Ugh.

One remaining imperfection: currently the timezone associated with `datetime` fields is
hard-coded to be UTC. Probably need to change that at *some* point. But also we should
always use UTC. Maybe the column name should end with `_utc`?

Not thrilled about having a column named `year` rather than `report_year` should that be
changed?

Closes #1200