Cache PUDL Package metadata, consolidate load modules
Created a function pudl.metadata.build_package() which caches the built PUDL metadata
Package for re-use, rather that reconstructing it all over the place in the code or
storing it as a global constant. Uses the functools.lru_cache decorator.

Consolidated the sqlite and parquet loaders into a single pudl.load module since
they're both so short.

Closes #1379
Closes #1380