Replace pudl.constants.COLUMN_DTYPES w/ Pydantic types
Rather than storing two separate sources of column/field type information, derive all of
the pandas data types from the now complete(-ish?) Pydantic metadata. Mostly this is
implemented by a small helper function pudl.metadata.fields.get_pandas_dtypes() which
refers to the canonical PUDL field metadata (and data group overrides) by default, but
which can also use independent field type dictionaries, so it can be used to enforce
types on analysis tables with an ad-hoc mix of canonical PUDL fields and other columns
that pertain to that analytical context.

One (temporary) change that was required to make this work is that all of the _year
fields are now being typed as integers, rather than years, to avoid having them be
converted into nanoseconds after Jan 1st, 1970.  When we try to handle the year, month,
and day resolutions of reporting in the harvesting, and merging dataframes of different
frequencies together, we'll want to undo this, but for now it maintains the existing
behavior and doesn't break anything, and it was a small change.

Closes #1270