Replace get_pudl_dtypes() with df.convert_dtypes()
It turns out that in most if not all of the cases where we are applying the canonical
PUDL dtypes, what we're actually doing is forcing all non-floats to use nullable Pandas
dtypes... and the built-in df.convert_dtypes() method does this.

I've hunted down and replaced all usage ofget_pudl_dtypes() with convert_dtypes() which
should avoid the need for our own typing infrastructure in almost all cases.

Excluding the conversion of floating point values is necessary since the pandas nullable
floats don't work with all of the Numpy functions.

There may be other instances (e.g. boolean columns with NA values that we use as masks)
where this doesn't work either... but haven't come across any just yet.