Use nullable string type for county names in planning areas.
Also looked into why there are a few more records showing up in the generation_fuel and
generation_fuel_nuclear tables now... It's not clear, but they look like valid records,
and they are all from 2001.

It seems that this change happened after we released PUDL v0.5.0 and before I started
messing around with the data types though -- the "additional" records were present in a
copy of the PUDL DB that I had from 2021-11-28, but were not there on 2021-11-11.