Exclude stochastic lines from coverage of timeseries_cleaning
There are several lines of code in the pudl/analysis/timeseries_cleaning.py module that
are hit or miss when it comes to testing, because the input data for the tests on that
module is stochastically generated. This results in small, meaningless changes in
coverage that prevent the CI from coming up all green sometimes. I'm trying to exclude
them from the coverage report as they come up to avoid this problem.