Fix breakage resulting from dask v2022.4.2
Dask introduced breaking changes to `dd.read_parquet()` in v 2022.4.2:
https://docs.dask.org/en/stable/changelog.html#v2022-04-2

However, we didn't catch this when it happened because it's only a
problem when there's more than one row-group. So now I'm processing
2019-2020 data for both ID and ME (two of the smallest states).  I was
already processing 4 state-years when testing the parallelized CEMS ETL
so this isn't any more data that we were pulling before.

Also fixed a bug that the parallelized ETL test introduced, where it was
writing the partitioned output to the same directory as the other CEMS
output, which might have been creating weird behavior.

I put more stringent bounds around the Dask versions in `setup.py` so
we'll at least be notified by @dependabot if a version changes now.

Moved the CI environment file under tests, since it's more related to
them than "devtools" and in preparation for configuring `tox-conda` to
use conda environments to (partly) manage the tox builds:

https://github.com/tox-dev/tox-conda