Move fuel_group_code to energy_sources_eia coding table
The fuel_receipts_costs_eia923 table reports an energy group, which must be one of coal,
petroleum, natural_gas, petroleum_coke, or other_gas. This is the category that the
fuels show up under in the EIA's Electric Power Monthly (EPM). It's entirely determined
by the value of `energy_source_code`, and so can be stored as one of the columns in the
`energy_sources_eia` coding table.

I've renamed it `fuel_group_eiaepm` to give a better indication of where it comes from,
as there are a number of other very generic energy categorizations in that table
already.