Pass the generators_entity_eia table to the EPACEMS crosswalk dquote> The epacems crosswalk file does not have a data field. It does, however, use foreign keys from eia tables that are restricted by date when the ETL is run for a subset of years. Without this addition, the integration tests will fail because the tests, based on the fast etl (aka one year of data), formulate entity tables (used to map foreign keys onto the crosswalk) that contain less data than the crosswalk (based on all years). One solution is to restrict the crosswalk data to only show the plant and generator ids from the selected subset of eia data. That's what I've implemented here.
This isn't ideal, because there are some cases in which the crosswalk has erronious generator_id data. I fixed this in the previous commit (before selecting only the plant-gen records that show up in the eia subset). If I hadn't checked these generator records they would have been excluded from the crosswalk for not matching. By only including the values that are inherently in the eia subset we guarantee to only get the data pertaining to the years we want BUT we run the risk of excluding data that was erroniously reported (and could be fixed) without warning.