Fix fix to helper function
The remove_leading_zeros_from_numeric_strings() function was faster than the old one, but there were some issues with it. First, there was a capture group in the regex causing it to throw an error. I removed the parenthesis from the regex to fix this.

Second, there was an error grabbing string rows from a column with NA values. I added a df[col_name].isna() to the specifications and the warning went away.