Update EPA-EIA crosswalk and add to CEMS transform
First, I updated the crosswalk from three tables to one. The previous tables figured that there was a 1:1 relationship between plant_id_epa and plant_id_eia. All the fields are interconnected, and it makes most sense for them to be in one table .I also added in the boiler_id_eia and generator_id_eia fields for mapping onto more granular eia data.

Next, I fixed some of the column names in CEMS. What we previously called  was a close but not perfect approximation of ORISPL code that the crosswalk intended to fix. I changed  to . There were also two unit columns in the CEMS data, one called  and one called .  referred to the smokestack units and  referred to...well...I still don't know. So I dropped the  column and renamed  to . Lastly, I removed the  column because it didn't seem to refer to anything useful.

After editing the column names, I updated the metadata, etl functions, and tests to reflect the changes.

The CEMS transform module was relying on the old  column to correct timezones in the data. While the  values were in the right timezones, it still wasn't technically the right , so I pulled the crosswalk into the CEMS transform module to replace the old  column (now ) with the correct . I did this by filling in the long empty harmonize_eia_epa_orispl() function. I also removed the add_facility_id_unit_id_epa() function because I removed those columns from the CEMS data.