Add boiler id to crosswalk and address gaps in plant_id_eia
Boiler id was commented out of the crosswalk metadata so I added it back in!

Adding in boiler id created some issues with foreign keys for the testing environment that only tests one year of data. I had to implement the same hack as for generator_id: check to see if the settings indicate all years of data and if not filter the crosswalk based on the boiler_id values in the boilers_entity_eia table. I also had to feed the boiler_entity table to all the same functions that the generator_entity table is fed to for the same reason.

Previously I had merged in the crosswalk plant_id_eia values and called that one of the crosswalk primary keys. Unfortunately the crosswalk still has some gaps and not all of the plant_id_epa values in cems have been mapped. That led to NA plant_id_eia values which isn't allowed for primary keys. Instead of deleting the plant_id_epa field in favor of the plant_id_eia field I kept both and called plant_id_epa the primary key.

This caused some other issues with the fix_up_dates() function that was relying on the plant_id_eia column to map timezones. To satisfy this function I created a temporary field called plant_id_combined that filled the gaps of plant_id_eia with plant_id_epa for use in timezone calculation. Technically there is no difference in timezone between the epa and eia ids (I checked) but this process is more accurate (and maybe something will change in future years). I later remove the plant_id_combined column.

Lastly, I had to update the tests because I added the plant_id_epa column back. Now there are 19 instead of 17 columns. I also added plant_id_epa back to the crosswalk schema.