Remove glue wrapper function
Previously I designed the glue epacems_unitid_to_eia_plant_crosswalk module with an extract function, a transform function, and a wrapper function that didn't do anything besides call the extract and transform. I removed that wrapper function and now call the extract and transform functions in the etl.py module. This is more in line with how the other BLAh_etl() functions work in that module anyways.