Update crosswalk module name
Was epacems_unitid_eia_plant_crosswalk but now is epacamd_eia_crosswalk. We no longer use the unitid column name and there are also other fields besides unit and plant. The data also pertains to CAMD more broadly, not just CEMS.

We still need to update the zenodo archiver and scraper and then update the old name in the pudl extractor.