Implement drop_invalid_rows() for fuel_ferc1 table
Create a fuel_ferc1 table specific drop_invalid_rows method, which both makes use of the
parameterized method inherited from the AbstractTableTransformer, and adds a separate
method for identifying rows which we believe to be plant totals, rather than rows that
pertain to individual fuels.

This results in dropping more than 2/3 of all the records in the fuel_ferc1 table.

I also made some name changes, trying to be a little more standard and (hopefully)
informative:

* remove_invalid_rows => drop_invalid_rows, since there were already several methods
  defined in the FERC 1 table transformers that do similar things in more specific
  circumstances, and they were named drop_*_cols or drop_*_rows.
* cols_to_check => required_valid_cols and cols_to_not_check => allowed_invalid_cols to
  provide some indication as to what is being checked for (validity / invalidity) in
  relation to the other transform parameter (invalid_values)
* DropInvalidRows => InvalidRows to follow the convention of the other TransformParams
  which are nouns describing their contents, while the methods & functions that they
  parameterize are verbs/actions.
* Moved the documentation of the InvalidRows parameters into the class definition.