Aggregate data_maturity in gfn_eia923, update EIA ETL debugging Notebook
Fix the aggregation of the `data_maturity` column in the transform step
for the `generation_fuel_nuclear_eia923` table, so the aggregated
records we create (to fix some missing primary key values) get the
`data_maturity` value of their input records and we don't have NA values
messing things up downstream.

Also update the EIA ETL Debugging notebook to reflect the current state
of the system on `dev`.

Closes #1914