Fix dataframe time column period formatting
This is a (small) hack that works around changes in how pandas deals
with casting Series to `datetime64` types have units larger than
seconds. We depended on the previous behavior. Not sure if it's
something that will get fixed in pandas, but I made this issue and it's
a breaking change, so hopefully:

https://github.com/pandas-dev/pandas/issues/48574