Add a simple TableTransformer class to unit tests.
Just to make sure all of the parts can be used independent of any
specific data source / collection of tables, and that the various parts
of the transform system work together in a simple arrangement.

Main issue that came up in this was the need to make sure all
TransformParams classes get reasonable default values when they aren't
specified in the TableTransformParams dictionary. Required some minor
modifications of the classes.