Add code to fix ferc license and headers
FERC license numbers are sometimes in the plant name and sometimes in the footnotes. This code labels row types as total, note, header, or NA and extracts the relevant information from each type of row and puts it in the right place. Headers get filled into a plant_type_2 column (for now), ferc license numbers get put in a ferc license column, and footnotes are moved to into a new column and the same row as the record they refer to.

This code also adds some hydro fuel types based on whether hydro is in the plant name or if there is a ferc license id for that record (only hydro records have ferc ids).

This code also fixes net_generation units which are actually reported in kwh and need to be converted to mwh. It also removes some old column names from the metadata (like plant_name_clean) and renames some (like ferc1_license_id to license_id_ferc1). It also corrects some of the plant rename values from before. Specifically some columns that were categorized as opex when they should have been capex.