Allow pre-processing to rename instant/duration XBRL cols
FERC is not always consistent in its naming of corresponding columns
that appear in both the instant & duration tables. This commit adds
a renaming step to the default FERC1 process_instant_xbrl() and
process_duration_xbrl() methods, that rename them early in the data
transformations. This transform is parameterized using the same
RenameColumns parameter and rename_columns() methods that we're using
elsewhere.