Add hydro to the PLANT_TYPE_CATEGORIES param
I added hydro and storage to the dict of PLANT_TYPE_CATEGORIES so it could be run on the small plants table. I kept the hydro table seperate because there was one instance, the word conventional, that, in the case of steam should be categorized as NA and in the case of hydro should be categorized as hydro. The hydro category in the PLANT_TYPE_CATEGORIES does not contain the phrase conventional whereas the hydro category in the PLANT_TYPE_CATEGORIES_HYDRO does. It also contains all the extra run-of-river categories.

I also alphabetized some of the plant type dicts and added the category name to the list of fuels so that it can be run more than once.