Add XBRL taxonomy metadata extraction integration test
I added this unit test while debugging a bunch of glue test failures,
which were somehow being triggered by the addition of the XBRL taxonomy
metadata. Whatever it's reading is from the XBRL metadata in the glue
tests... it's unable to parse as JSON, resulting in a
NotImplementedError.

This problem does *not* arise with the new XBRL metadata integration
test, so it seems there's something weird going on with the glue test
environment in particular.

Another strange thing I noticed is that even if you run the tests with
--live-dbs the glue tests run the ETL anyway! Not sure why this would be
happening. Running

```
pytest --live-dbs test/integration/etl_test.py::test_ferc1_etl
```

also doesn't exhibit this behavior, and it finishes in just a few
seconds.

I split the FERC 1 ETL test into two tests, one for DBF and one for XBRL,
and renamed them:

  * test_ferc1_dbf2sqlite()
  * test_ferc1_xbrl2sqlite()

Rather than "etl" since they're only doing the conversions of the raw data. This lets
you just test one or the other too, which is more convenient when the problem you're
working on is just on one side.