Use default io manager for interim tables and sqlite io manager for final tables [no ci]
This version of the EIA ETL uses the fs_io_manager for the interim tables and the SQLiteIOManager
for the tables that are in our metadata. This way we don't have to worry about adding interim
table metadata right now. This version produces a sqlite constraint error when adding the
ownership_eia860.owner_zip_code field. All of the values are 5 digits so I'm not sure what's
causing the error.