fix fact vs factoid name from metadata
the plant in service table's normalize metadata override replaces
the fact name and uses *both* the renamed fact and the original
name. not sure if this is strictly necessary, but for now I will
keep this in. i'm currently looking into why this is happeneing
and if/how to remove the additional complication.