Update table descriptions. Remove obsolete static table.
The static ferc_depreciation_lines table is no longer relevant, with the
addition of the XBRL metadata and new transforms for the accumulated
depreciation table, so it has been removed.

Also updated the table descriptions of several new FERC tables to
indicate what schedule they come from in the Form 1 to make it easier
for people to make the connection between PUDL data and FERC info.