Generalize unstack balances to work with any PKs
The electric_plant_depreciation_changes_ferc1 table has starting/ending
balances in the instant table, but it also has non-standard primary keys
in that table, which the existing unstack method couldn't accommodate.

Now instead of hard-coding the primary keys in the unstack function,
they are passed by the method wrapper, based on the transformer class.