Since we moved the plant part list and the ferc-to-eia record linkage to PUDL from the rmi-ferc1-eia repo, I also needed to move the notebook that created the tools for the training data/overrides.
I moved the notebook over to this repo and renamed it from make_overrides.ipynb to make_ferc1_eia_training_data.ipynb to be more specific.

I also moved all the functions it references from the rmi-ferc1-eia module create_override_spreadsheets.py to the pudl.analysis.ferc1_eia.py module. I wasn't sure exactly where to put them, so I added them at the end of this module. These functions are a tad hacky, but they aren't used in any of the main PUDL processing so I didn't invest much time in cleaning them up.

I did add another function called already_in_training that allows you to see whether any values are or are not already in the training from any given spreadsheet you're working on. This is nice if you aren't sure whether or how much of the data has already been integrated into the official training or not.

I also updated the functions to look for the pudl versions of the ppl and ferc1_eia tables.