Migrate all clean EIA-861 tables into database.
* Enabled all EIA-861 resource definitions.
* Added `data_maturity` column to all EIA-861 tables.
* Fixed issues in all EIA-861 transforms related to `data_maturity`
* Switched to writing all EIA-861 (clean but not not normalized) tables into the DB.
* Removed the "interim" EIA-861 ETL from `PudlTabl` output object.
* Changed all EIA-861 `PudlTabl` output methods to read from the DB.

Note: only a few of the EIA-861 tables have primary keys defined, so these are
fast-and-loose tables. "Clean" but not normalized, and very lightly constrained. We
should probably label most of the with the `clean_` prefix and document their status
very clearly so people know. But they're available in the database now instead of hidden
behind a complicated software interface!!!!