Update naming protocol for current training data (sometimes was calling it old training and sometimes current training) so that it's always current training.
Also, change how importlib is reading in the training data and null overrides. Previously it was using importlib.resources.open_text() but now it is using importlib.resources.files(). open_text() would not allow multiple versions of the spreadsheet to be open at the same time. It appeared to be more than just a path. We need more than one version of the training data to be readable because we use it to validate as well as add training data to.