WIP: refactoring of the foxpro extractor
Collapsed some types for better clarity, extracted TableSchema as an
independent type; the code is starting to look somewhat reasonable.

dbf2sqlite still needs to be alterated to work with this new API, and
to generalize the FercExtractor logic into separate classes
(FoxProExtractor), that will exist in addition to the smart datastore
that can load dfs and create sql schemas.

This is where the dataset specific logic, such as:
1. PK/FK constraints on the tables
2. table-specific transformations that need to be run before
   data is injected into sqlite
3. late transformation that operate on the sqlite contents