Split up denormalized EIA assets. Add aggregating asset factory
For organizational purposes, split the generic EIA (entity) assets,
EIA-860 (only ownership), and EIA-923 assets into separate modules, as
the EIA-923 assets are going to be a bit more complex.

Create an asset factory for the generation_eia923 table(s) that
aggregates to a desired temporal frequency, before using the same
process to denormalize the resulting output dataframe.