Now that utility_id_pudl is in the ppe, we don't need to merge with the utils_eia860 table in the validation step to get utility_id_eia and then utility_id_pudl. So remove the merge with utils_eia860 step from the validate_override_fixes function. Also update the tests and notebooks to reflect this.
Also changed the name of some of the override sheet columns: utility_id_pudl to utility_id_pudl_ferc1 because it originates from a ferc table and plant_id_pudl to plant_id_pudl_ferc1 because it originates from a ferc1 table. Previously the RENAME_COLS_FERC1_EIA dict was looking for columns of that name that didn't exist. Instead of looking for plant_id_pudl_ferc1 and changing it to plant_id_pudl, this change does the reverse. Adding the _ferc1 suffix helps the user know where this value comes from which is important during the mapping process when matches are occasionally accross different plants or utilities. It's important to know when that is the case.

These changes also combine all the test cases for the validate_override_fixes function into one. It was a little tricky to find test cases that fit all the parameters and would break in the right way (ex: a utility that's from 2020, has some EIA records in the training data already and some FERC records not in the training data -- considering we've built the training data by scouring each utility in entirety.) It's possible that this would not work in future years (because the data are still dependent on the fast tests and need to be updated with new years of data that fit those same conditions).

It's not ideal to have one function that does all these tests, but it's ok for now.