Rename check_table_calculations; fillna() in correction calcs
I renamed check_table_calculations / CheckTableCalculations to
reconcile_table_calculations / ReconcileTableCalculations because
we're not just checking the values now -- we're altering the data in
many cases, to include correction records, and accounting reconciliation
has a more active "make these things match" connotation.

Also now using `fillna()` when generating the correction records, so
that if we're correcting a calculated value that's NA we actually get
a real correction (we're treating NA values as 0.0 in the aggregations
already so this is internally self-consistent).