Fix typo in explode_tables() type hint.
For some reason the tests don't seem to be hitting the functions that
are used in generating the exploded tables, so I'm curious if they're
actually getting run as part of the ETL -- they do seem to be included
in the DAG as displayed by dagit locally.

Fixing the hint typo, adding some additional type hints, made use of
exploded_tables consistent in the asset definitions / helper functions.
Hoping that in the re-run of the CI these functions get hit now.

This type hint typo is also what kept the docs build from suceeding.