Manage duplicate PKs in EIA-861 table transforms
The `sales_eia861` and `demand_response_eia861` tables each have a
handful of duplicate primary keys due to NA values in the
`balancing_authority_code_eia` column. Quantify and log the extent of
this problem, and consolidate the data in the duplicated records if they
constitute less than 0.5% of all records in the table.

This check would also have caught the incorrect primary key columns
reported in #2636 and fixed in #2637.

Because there were so few duplicate records, I decided to just
consolidate them all (with a hard limit on the fraction of records that
could be consolidated) rather that requiring that the only duplication
be due to the BA Code column.

Closes #2638