Deal with complexities of passthrough calculations
In some cases redirecting a calculation to point at the same value
reported in a more granular table results in a node having more than one
parent.  This commit prunes those passthrough nodes after the tree is
constructed, since they don't provide any additional structural
information.

In addition, when the referenced fact has a weight of -1.0 in the actual
calculation (the redirect always has a weight of 1.0) you can end up
with two conflicting definitions of the same node in the tree when it's
being constructed -- a conflict that would be resolved later when the
passthrough node is pruned -- but we don't know that at the time the
tree is being constructed.  As a hack, when we find this kind of weight
conflict, we always use a weight of -1.0 (since a weight of 1.0 is a
no-op, and the fact that there's a conflict means one of them is -1.0)

Also fixes some apparently incorrect calculation fixes.