fix redefine_calculations_with_components_out_of_explosion to actually filter things out!
i had made a short cut bc this method we will throw away. i was just using the
saved processed exploded calc component table to detemine if any of the
calculations had out-of-explosion elements. but i had already filtered those
out in the processed calc component table. so nothing was getting filtered.

it is mega copy/pasty rn but it will be gone soooo sooooon