Allow a mix of Zenodo sandbox & production DOIs
Okay I did this off the clock since it has been driving me a little bit nuts.

Historically we've required that all Zenodo DOIs in the datastore come either
from the Sandbox or the Production server, which makes testing a single new
archive on its own a hassle, and adds complexity across the whole application
with switches for sandbox vs. not-sandbox data sources.

This commit removes this requirement, and allows a mix of sandbox and production
DOIs to be used in development.

I also removed some very sparse documentation about how to create an archive in
the Datastore by hand, which I think was very old and probably no longer supported
and certainly not being tested, since it seemed likely to confuse and frustrate
anyone who actually tried to do it.

There's a unit test which checks that all DOIs are production, rather than sandbox
to make it difficult to accidentally check in code that refers to unofficial
input data.