Simplify code linting & formating; use ruff
ruff is a Python linter written in Rust. It's literally 1000x as fast as flake8,
and re-implements the functionality of a bajillion flake8 plugins and more. Using
ruff means our linting setup is both much simpler and much faster.

Changes include:
- Moved all remaining tool configuration from tox.ini to pyproject.toml
- Replaced bandit with ruff
- Replaced pyupgrade with ruff
- Replaced all flake8 plugins with ruff
- Updated tox testenvs to reflect use of ruff
- Update .pre-commit-config.yaml to reflect use of ruff
- Removed other linter dependencies from dev & test environments
- Added ruff as a dependency in the test & dev environments
- Fixed a few new issues identified by ruff
- Removed the rstcheck linter, since there's no ruff support for it yet and it's
  been causing dependency issues in our Docker builds.
- Removed the --sign option from twine since PyPI no longer supports it.
- Updated to build v1.0.0

Overall I tried to keep the linting rules applied in this PR as close to the
old ones that we had specified with flake8 and friends. We can add more rules from
the ruff repertoire incrementally.