Extract more data from FERC XBRLs and handle that new data in ETL (#2821)
* Update to use new version of ferc-xbrl-extractor

* Fix issues arising from stricter typing used in pandas 2.1

* Use integer transmission circuits.

* Remove obsolete references to ferc1_schema tests.

* Make new extractor compatible with 2021 data

The new extractor added some data to the 2021 XBRL archives. This caused some integration and validation test fails. I added some plants to the pudl_id mapping spreadsheet, all of which are considered totals. I.e., not real plants, but we're mapping them for the sake of giving them an ID (they are not connected to EIA records). Because this is how we treat other total records reported to FERC1.

This also updates the way that values were assigned to a slice of the ferc1_eia_train output spreadsheets. NA values were causing an issue, so I had to change how the values were being converted.

This also updates the test_minmax_rows test to reflect the new rows in the 2021 data.


* Add a few plants to pudl_id_mapping

Totally new:

* 18012: pjm interconnection, llc / total
* 18013: new york state electric & gas corporation / see footnote
* 18014: southwest power pool, inc. / total
* 18015: public service company of colorado / community solar gardens
* 18016: the empire district electric company / n/a
  each & 73 units at 2.52 mw each)
* 18017: wisconsin electric power company / see footnote
* 18018: upper michigan energy resources company (pudl determined) / total
* 18019: new york transco, llc / total
* 18020: wilderness line holdings, llc / total
* 18021: mt. carmel public utility co / total

Mapped to existing PUDL ID:

* 8671: pacific gas & electric company, small hydroelectric generating plants
* 15000: idaho power company / hydro
* 15001: idaho power company / internal combustion
* 15068: public service company of colorado / conventional hydro
* 12926: midamerican energy company / ida grove ii wind farm (8 units at 2.3 mw
* 1287: alaska electric light and power company / salmon creek hyrdo

Note the misspelling of the plant name in 1287.

Changed:

* 15031: mt. carmel public utility co / not applicable -> ameren
  illinois company / not applicable

  This one had a mismatch between utility_id_ferc 222, which corresponds
  to Ameren, not Mt. Carmel (397).

* Update validation test expectations.

There are some missing data due to messy deduplication:
https://github.com/catalyst-cooperative/pudl/issues/2822

But we'll do the deduplication better in here:
https://github.com/catalyst-cooperative/pudl/issues/2899

---------

Co-authored-by: zschira <zach.schira@catalyst.coop>
Co-authored-by: Zane Selvans <zane.selvans@catalyst.coop>
Co-authored-by: Austen Sharpe <austensharpe@gmail.com>