Update pull_request_template.md to improve full ETL instructions (#3446)
* Update pull_request_template.md to improve full ETL instructions

The full ETL often can't be run locally. These edits 1) move the full ETL to be the last item in the checklist, reflecting the fact that it should be run after all other items are checked off and 2) highlights the particular GHA that can be used to run the full ETL on batch if it cannot be run on one's local device.

* Clarify validation tests

* Update distinction between minor and major changes