add two new correction records into plant_in_service table (#3525)
these two new records are a result of adding corrections when either
the reported or calculated value is null. this change did not add a
ton of extra records to a ton of tables because most of the tables
that get their calculations reconciled/corrected are not the orignal
ferc1 tables that are in the PudlTabl. These calculation-reconciled
tables are mostly new and are mostly not being tested in our test_minmax_rows
tests.