Post v2025.2.0 cleanup
- Remove the no longer relevant Parquet bucket checks from versioned
  release issue template.
- Bump a couple of significant versions noticed when updating conda
  feedstock
- Re-lock conda dependencies (including dagster v1.10.0)
- Update pre-commit hooks.

Since there's substantial Dagster update and I'm unable to materialize
all locally (for reasons I don't understand) I want to try running the
ETL in CI and see if the problem is me, or the new Dagster.  I tried
doing materialize all on dagster 1.9.13 and it also didn't work, so my
guess is it's me somehow.