Normalize and clean SEC 10-K company information tables
- Split the company information table out into two better normalized tables:
  core_sec10k__quarterly_company_information and
  core_sec10k__changelog_company_name
- Separate the processing into two separate steps, first a multi-asset
  that reshapes the raw data but makes minimal changes to the contents,
  and then one asset for each of the above tables to do the
  table-specific data cleaning.
- Start moving assets out of `analysis/pudl_models.py` into SEC 10-K
  extract, transform, and output modules using the same pattern as our
  other data sources.

With this commit, all of the assets in the `core_sec10k` group
materialize and the asset checks pass -- however there are a couple of
ENUM constraints that seem to be failing because the columns contain NA
values, but I think that's supposed to be allowed. These ENUM
constraints are currently commented out in `pudl/metadata/fields.py`.

I tried to put a number of the new field definitions into a data source
specific group definition, but PACKAGE_METADATA wasn't picking them up
there for some reason, and after a bit of debugging I gave up. Should
figure out what's going on here though.