Migrate vs_bounds validation tests to dbt (#4171)
* Modify dbt_helper to combine entries for single-quantile vs_bounds tests

* Add test macro for single-entry quantile checks

* dbt_helper: build single-entry quantile tests from test config variable

* Wipe out sample old-style quantile tests

* Convert comments to descriptions and correct handling of weight column

* Call dbt_expectations test correctly

* Permit weighted quantile macros to abide by row conditions

* Add macro to reduce toil while debugging vs_bounds tests

* Make failure output more legible

* Migrate vs_bounds for _out_eia__yearly_generators

* Migrate vs_bounds tests for table out_eia923__monthly_boiler_fuel

Reduced bounds on one quantile from 5.0 to 4.9 to account for minor shift on etl-fast, and annotated to investigate if it gets worse.

Demoted one quantile check to warning due to signficant shift on etl-fast, and annotated with relevant tracking issue for investigation.

* Migrated vs_bounds tests for table out_eia923__fuel_receipts_costs

Reduced bounds for one quantile from 18.0 to 17.5 and annotated for attention in case it drops again.

* Migrate vs_bounds tests for table out_eia923__generation_fuel_combined

Reduced bounds for two quantiles from 5.0 to 4.79 and from 5.5 to 5.49 and annotated for attention in case they drop again.

* Migrate vs_bounds tests for table out_eia923__yearly_boiler_fuel

Demoted one quantile check to warning due to significant drop in etl-fast. Annotated with reference to tracking issue.

* Fix null criteria in weighted quantile macro

* Reduce bounds on one quantile from 5.0 to 4.8 to pass etl-fast, and annotate for attention to investigate if it gets worse.

* Migrate vs_bounds tests for table out_eia__monthly_generators.

Reduced bounds on two quantile checks from 0.04 to 0.036 and from 1.65 to 1.489 account for minor shifts on etl-fast, and annotated to investigate if it gets worse.

Demoted one quantile check to warning due to signficant shift on etl-fast, and annotated with relevant tracking issue for investigation.

* Migrate vs_bounds tests for table out_eia__yearly_generators.

Reduced bounds for two quantile tests from 0.5 to 0.444 and from 1.65 to 1.63 to account for minor shifts in etl-fast, and annotated for attention to investigate if it gets worse.

Demote one quantile test to a warning due to major shifts in etl-fast, and annotated with tracking issue for further investigation.

* Add handling for model_name in migrate_tests. Add positioning for description in DbtTable.

* Migrate vs_bounds tests for table out_ferc1__yearly_steam_plants_fuel_by_plant_sched402.

Raise bounds for one quantile from 25.0 to 26.1 to pass etl-fast, and annotate for attention to investigate if it gets worse.

Add missing fuels to intermediate model.

* Migrate vs_bounds tests for table out_ferc1__yearly_steam_plants_fuel_sched402.

Raise bounds for one quantile test from 0.95 to 1.095 to account for minor shift in etl-fast, and annotate for attention to investigate if it gets worse.

Demote one quantile test to warning to account for major shift in etl-fast, and annotate with tracking ticket for further investigation.

* Rename debug macro

* Migrate vs_bounds tests for table out_ferc1__yearly_steam_plants_sched402.

Demoted 5 checks to warnings due to significant quantile shifts in etl-fast, and added tracking issues for further investigation.

Lightly-modified 2 quantiles due to minor shifts in etl-fast, and annotated for attention to investigate if they get worse.

* Fixed severity and decimal place mistakes

* Update unit tests for weighted_quantile and move to centered interpolation

* Clean up typos and logging

* Remove old references to vs_bounds

* [pre-commit.ci] auto fixes from pre-commit.com hooks

For more information, see https://pre-commit.ci

* Improve coverage for dbt_helper: test migrate_tests and _get_config

* Add vs_bounds to release notes. And better labeling in debug macro.

* Remove deprecated weighted quantile vs. self validation specs.

* Updated values for etl-full

* Add documentation for quantile debugging macro.

* Adapt Python copy of weighted_quantile to match the approach in dbt.

* [pre-commit.ci] auto fixes from pre-commit.com hooks

For more information, see https://pre-commit.ci

* Fix line lengths in debug macro howto

* Fix line lengths in testing docs

* Add debug macro docs to hidden testing toc (as a stopgap until we have real dbt docs where it can more naturally live)

---------

Co-authored-by: Kathryn Mazaitis <kathryn.mazaitis@catalyst.coop>
Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: Zane Selvans <zane.selvans@catalyst.coop>