Remove no_null_rows tests
In an attempt to cut down memory usage in our CI, I went hunting for
tests that could be migrated or deprecated and found that only the
MCOE tables were being checked with pudl.validate.no_null_rows(), so
I moved those tests into asset checks on the MCOE assets.

In the process I found that there was some PudlTabl cruft in the
integration and validation tests that looked like it was also
unnecessarily reading a large dataframe out of pudl.sqlite.